Do not try to convert results to numbers
Below is tested on Org 9.4.6-544-gc5573b

Before this commit:

    #+begin_src nim
    echo "-0"
    #+end_src
    #+RESULTS:
    : 0

After this commit:

    #+begin_src nim
    echo "-0"
    #+end_src
    #+RESULTS:
    : -0

There was a bug in `org-babel--string-to-number` which got fixed
recently in
https://code.orgmode.org/bzg/org-mode/commit/4c934424d6e45daa30d2e5a2b3217181216d4e37
. This piece of `ob-nim` code was affected because `org-babel-read`
was using `org-babel--string-to-number` internally.

Reviewing the code in `org-babel-read` and also looking at
`ob-python.el`, it looks like that function isn't required here. So
using just `results` directly.

This commit also adds a test that proves this fix.